+++ wry/ADNMappingProvider.h
@@ -14,5 +14,9 @@
 + (RWJSONMapping *)userDescriptionMapping;
 + (RWJSONMapping *)postMapping;
 + (RWJSONMapping *)fileMapping;
++ (RWJSONMapping *)messageMapping;
++ (RWJSONMapping *)channelMapping;
++ (RWJSONMapping *)annotationMapping;
++ (RWJSONMapping *)accessControlListMapping;
 
 @end