+++ wry/PmCommand.h
@@ -1,0 +1,13 @@
+//
+//  PmCommand.h
+//  wry
+//
+//  Created by Rob Warner on 05/07/13.
+//  Copyright (c) 2013 Rob Warner. All rights reserved.
+//
+
+#import "WryCommand.h"
+
+@interface PmCommand : NSObject <WryCommand>
+
+@end